refactor: reorganize component imports and remove unused components
