+++ app/components/icons/wordmark.tsx
